Combine workspace_comment_render_svg.js with workspace_comment_svg.js
