+++ tests/playground.html
@@ -71,7 +71,6 @@
 <script>
 // Custom requires for the playground.
 goog.require('Blockly.WorkspaceCommentSvg');
-goog.require('Blockly.WorkspaceCommentSvg.render');
 </script>
 <script>
 'use strict';
