+++ tests/playgrounds/advanced_playground.html
@@ -77,7 +77,6 @@ goog.require('Blockly.Themes.Zelos');
 
 // Other.
 goog.require('Blockly.WorkspaceCommentSvg');
-goog.require('Blockly.WorkspaceCommentSvg.render');
 </script>
 <script>
 'use strict';
